moved embedding and reconstruct into separate files
